remove peristance and file migrator from robot servr
