+++ robot-server/robot_server/persistence/_migrations/v04_to_v05.py
@@ -9,7 +9,8 @@ Summary of changes from schema 4:
 from contextlib import ExitStack
 from pathlib import Path
 
-from .._folder_migrator import Migration
+from server_utils.persistence.folder_migrator import Migration
+
 from ..database import sql_engine_ctx
 from ..file_and_directory_names import DB_FILE
 from ..tables import schema_05
