+++ robot-server/robot_server/persistence/_migrations/v05_to_v06.py
@@ -22,7 +22,8 @@ from pathlib import Path
 
 import sqlalchemy
 
-from .._folder_migrator import Migration
+from server_utils.persistence.folder_migrator import Migration
+
 from ..database import sql_engine_ctx, sqlite_rowid
 from ..file_and_directory_names import (
     DATA_FILES_DIRECTORY,
