+++ robot-server/robot_server/persistence/_migrations/v07_to_v08.py
@@ -12,7 +12,8 @@ from pathlib import Path
 
 import sqlalchemy
 
-from .._folder_migrator import Migration
+from server_utils.persistence.folder_migrator import Migration
+
 from ..database import sql_engine_ctx
 from ..file_and_directory_names import (
     DB_FILE,
