+++ robot-server/robot_server/persistence/_migrations/v09_to_v10.py
@@ -21,8 +21,8 @@ from opentrons.protocols.api_support.deck_type import (
 from opentrons.types import DeckSlotName
 from opentrons_shared_data.deck import load as load_deck
 from opentrons_shared_data.deck.types import DeckDefinitionV5
+from server_utils.persistence.folder_migrator import Migration
 
-from .._folder_migrator import Migration
 from ._util import copy_contents
 from robot_server.persistence.database import sql_engine_ctx
 from robot_server.persistence.file_and_directory_names import DB_FILE
