+++ robot-server/robot_server/persistence/_migrations/v13_to_v14.py
@@ -7,7 +7,8 @@ This schema version adds the camera_capture_image_settings_table which contains 
 
 from pathlib import Path
 
-from .._folder_migrator import Migration
+from server_utils.persistence.folder_migrator import Migration
+
 from ..database import sql_engine_ctx
 from ..file_and_directory_names import DB_FILE
 from ..tables import schema_14
